+++ src/ModularPipelines/Engine/ServiceProviderInitializer.cs
@@ -1,4 +1,4 @@
-using TomLonghurst.Microsoft.Extensions.DependencyInjection.ServiceInitialization.Extensions;
+using Initialization.Microsoft.Extensions.DependencyInjection.ServiceInitialization.Extensions;
 
 namespace ModularPipelines.Engine;